Register ServiceBusSender
