+++ source/dotnet/wholesale-api/Events/Communication/Communication/IIntegrationEventProvider.cs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Core.Messaging.Communication.Internal;
+
 namespace Energinet.DataHub.Core.Messaging.Communication;
 
 /// <summary>
