+++ source/dotnet/wholesale-api/Events/Communication/Communication/Internal/IServiceBusMessageFactory.cs
@@ -14,7 +14,7 @@
 
 using Azure.Messaging.ServiceBus;
 
-namespace Energinet.DataHub.Core.Messaging.Communication;
+namespace Energinet.DataHub.Core.Messaging.Communication.Internal;
 
 public interface IServiceBusMessageFactory
 {
