+++ source/dotnet/wholesale-api/Events/Communication/Communication/Internal/OutboxSenderTrigger.cs
@@ -1,0 +1,37 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Microsoft.Extensions.Logging;
+
+namespace Energinet.DataHub.Core.Messaging.Communication.Internal;
+
+/// <summary>
+/// The sender runs as a background service
+/// </summary>
+public class OutboxSenderTrigger : RepeatingTrigger<IOutboxSender>
+{
+    private const int DelayInSecondsBeforeNextExecution = 10;
+
+    public OutboxSenderTrigger(
+        IServiceProvider serviceProvider,
+        ILogger<OutboxSenderTrigger> logger)
+        : base(serviceProvider, logger, TimeSpan.FromSeconds(DelayInSecondsBeforeNextExecution))
+    {
+    }
+
+    protected override async Task ExecuteAsync(IOutboxSender outboxSender)
+    {
+        await outboxSender.SendAsync().ConfigureAwait(false);
+    }
+}
