+++ source/dotnet/wholesale-api/Events/Communication/Communication/Registration.cs
@@ -12,18 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Core.Messaging.Communication.Internal;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Energinet.DataHub.Core.Messaging.Communication;
 
 public static class Registration
 {
-    public static IServiceCollection AddCommunication<TIntegrationEventProvider>(this IServiceCollection services)
+    public static IServiceCollection AddCommunication<TIntegrationEventProvider>(
+        this IServiceCollection services,
+        string serviceBusIntegrationEventWriteConnectionString,
+        string integrationEventTopicName)
         where TIntegrationEventProvider : class, IIntegrationEventProvider
     {
         services.AddHostedService<OutboxSenderTrigger>();
 
         services.AddScoped<IIntegrationEventProvider, TIntegrationEventProvider>();
+        services.AddSingleton<IServiceBusSenderProvider>(
+            _ => new ServiceBusSenderProvider(serviceBusIntegrationEventWriteConnectionString, integrationEventTopicName));
         services.AddScoped<IOutboxSender, OutboxSender>();
         services.AddScoped<IServiceBusMessageFactory, ServiceBusMessageFactory>();
