+++ source/dotnet/wholesale-api/Events/Events.Application/CalculationResultPublishing/ICalculationResultIntegrationEventFactory.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Core.Messaging.Communication;
+using Energinet.DataHub.Core.Messaging.Communication.Internal;
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
 
 namespace Energinet.DataHub.Wholesale.Events.Application.CalculationResultPublishing;
