+++ source/dotnet/wholesale-api/WebApi/Configuration/Modules/EventsRegistration.cs
@@ -14,6 +14,7 @@
 
 using Energinet.DataHub.Core.JsonSerialization;
 using Energinet.DataHub.Core.Messaging.Communication;
+using Energinet.DataHub.Core.Messaging.Communication.Internal;
 using Energinet.DataHub.Wholesale.Events.Application;
 using Energinet.DataHub.Wholesale.Events.Application.CalculationResultPublishing;
 using Energinet.DataHub.Wholesale.Events.Application.CompletedBatches;
@@ -24,7 +25,6 @@ using Energinet.DataHub.Wholesale.Events.Infrastructure.EventPublishers;
 using Energinet.DataHub.Wholesale.Events.Infrastructure.IntegrationEvents.Factories;
 using Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence;
 using Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence.Batches;
-using Energinet.DataHub.Wholesale.Events.Infrastructure.ServiceBus;
 
 namespace Energinet.DataHub.Wholesale.WebApi.Configuration.Modules;
 
@@ -43,14 +43,13 @@ public static class EventsRegistration
         serviceCollection.AddScoped<ICompletedBatchRepository, CompletedBatchRepository>();
         serviceCollection.AddScoped<ICompletedBatchFactory, CompletedBatchFactory>();
         serviceCollection.AddScoped<IRegisterCompletedBatchesHandler, RegisterCompletedBatchesHandler>();
-        serviceCollection.AddIntegrationEventPublisher(serviceBusConnectionString, integrationEventTopicName);
 
         serviceCollection.AddScoped<ICalculationResultIntegrationEventFactory, CalculationResultIntegrationEventToIntegrationEventFactory>();
 
         serviceCollection.AddApplications();
         serviceCollection.AddInfrastructure();
 
-        serviceCollection.AddCommunication<IntegrationEventProvider>();
+        serviceCollection.AddCommunication<IntegrationEventProvider>(serviceBusConnectionString, integrationEventTopicName);
     }
 
     private static void AddApplications(this IServiceCollection services)
